backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
pandas
python-dotenv
jinja2
google-generativeai
